pyproject.toml
[tool.poetry]
name = "collaberr-backend"
version = "0.1.0"
description = ""
authors = ["Tony Lee <43350428+TonysHub@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"


[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
flake8 = "^6.0.0"
django-split-settings = "^1.2.0"
pillow = "^9.5.0"
black = "^23.3.0"
pytest = "^7.3.2"
pytest-xdist = "^3.3.1"
pytest-django = "^4.5.2"
model-bakery = "^1.12.0"
psycopg2 = "^2.9.6"
 

[tool.poetry.group.dev.dependencies]
colorlog = "^6.7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
